move tx building to v4 builder
